Supporting JWT
Also did a bit of clean up of package names.
Also introduced an interface for the services to implmenent.
This is done so we can update the interface definition when adding new
rest methods, and that will give us an not implemented error in the
actual handler with the business code